backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
elasticsearch[async]==7.17.9
openai==1.10.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
aiofiles>=23.0.0
